Refactor code and add support for dotbot defaults
This PR adds support for defaults to each of the directives supported by
dotbot-brew. The defaults are booleans for 'stdin', 'stdout', 'stderr'
which control whether the underlying commands get any of these handles
or are redirected to /dev/null.

Furthermore this PR refactors code to improve code reuse throught the
whole Brew class.